+++ tpi/list_t.h
@@ -4,6 +4,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+
  
 typedef struct snodo{ //snodo es el nombre de la estructura
     double valor;
@@ -12,7 +13,7 @@ typedef struct snodo{ //snodo es el nombre de la estructura
  
 typedef tnodo *tpuntero; //Puntero al tipo de dato tnodo para no utilizar punteros de punteros
  
-void insert_InList(tpuntero *cabeza, int e);
+void insert_InList(tpuntero *cabeza, double e);
 void print_List (tpuntero cabeza);
 void delete_List (tpuntero *cabeza);
 void list_new(tpuntero *cabeza);
